rename: order, restaurant 폴더를 domain 폴더 하위로 옮기기 완료
