+++ src/main/java/com/unknown/deliveryserver/domain/order/entity/OrderItems.java
@@ -1,7 +1,6 @@
-package com.unknown.deliveryserver.restaurant.entity.menu;
+package com.unknown.deliveryserver.domain.order.entity;
 
 import com.unknown.deliveryserver.global.common.entity.BaseEntity;
-import com.unknown.deliveryserver.restaurant.entity.Restaurant;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
@@ -10,29 +9,23 @@ import lombok.*;
 import lombok.experimental.SuperBuilder;
 import org.hibernate.annotations.Comment;
 
-import java.math.BigDecimal;
-
 @Entity
 @Getter
 @SuperBuilder(toBuilder = true)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class Menu extends BaseEntity {
+public class OrderItems extends BaseEntity {
 
     @ManyToOne // 다대일 단방향 (외래키를 갖는 쪽이 주인 엔티티)
-    @JoinColumn(name = "store_id")
+    @JoinColumn(name = "order_id")
     @ToString.Exclude
-    private Restaurant restaurant;
+    private Order order;
 
-    @Comment("대표메뉴 여부")
-    @Column(name = "is_primary", columnDefinition = "BOOLEAN")
-    private String isPrimary;
-
-    @Comment("메뉴(음식) 이름")
+    @Comment("주문한 음식 이름")
     @Column(name = "food_name", columnDefinition = "VARCHAR(200)")
     private String foodName;
 
-    @Comment("메뉴(음식) 가격")
-    @Column(name = "price", columnDefinition = "DECIMAL(64, 3)")
-    private BigDecimal price;
+    @Comment("주문한 음식 수량")
+    @Column(name = "quantity", columnDefinition = "BIGINT")
+    private Long quantity;
 }
