+++ src/main/java/com/unknown/deliveryserver/domain/order/enumerated/OrderStatus.java
@@ -1,0 +1,4 @@
+package com.unknown.deliveryserver.domain.order.enumerated;
+
+public enum OrderStatus {
+}
