+++ src/main/java/com/unknown/deliveryserver/domain/restaurant/dao/RestaurantRepository.java
@@ -1,7 +1,7 @@
-package com.unknown.deliveryserver.restaurant.dao;
+package com.unknown.deliveryserver.domain.restaurant.dao;
 
 
-import com.unknown.deliveryserver.restaurant.entity.Restaurant;
+import com.unknown.deliveryserver.domain.restaurant.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
