+++ src/main/java/com/unknown/deliveryserver/domain/restaurant/entity/Restaurant.java
@@ -1,7 +1,7 @@
-package com.unknown.deliveryserver.restaurant.entity;
+package com.unknown.deliveryserver.domain.restaurant.entity;
 
 import com.unknown.deliveryserver.global.common.entity.BaseEntity;
-import com.unknown.deliveryserver.restaurant.dto.request.RestaurantRequest;
+import com.unknown.deliveryserver.domain.restaurant.dto.request.RestaurantRequest;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import lombok.AccessLevel;
