--- src/main/java/com/unknown/deliveryserver/order/enumerated/OrderStatus.java
@@ -1,4 +1,0 @@
-package com.unknown.deliveryserver.order.enumerated;
-
-public enum OrderStatus {
-}
